skills-reference/livekit-voice-agent/templates/pyproject.toml
[project]
name = "livekit-voice-agent"
version = "0.1.0"
description = "LiveKit voice agent with multi-agent workflows"
readme = "README.md"
requires-python = ">=3.9,<3.14"
dependencies = [
    # Core LiveKit Agents framework
    "livekit-agents>=1.3.3",

    # LLM Provider (OpenAI)
    "livekit-plugins-openai",

    # Speech-to-Text Provider (Deepgram)
    "livekit-plugins-deepgram",

    # Voice Activity Detection (Silero)
    "livekit-plugins-silero",

    # Environment variables
    "python-dotenv",
]

[project.optional-dependencies]
# Development dependencies
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
]

# Alternative LLM providers
anthropic = ["livekit-plugins-anthropic"]
groq = ["livekit-plugins-groq"]

# Alternative STT providers
assemblyai = ["livekit-plugins-assemblyai"]
google = ["livekit-plugins-google"]

# Alternative TTS providers
elevenlabs = ["livekit-plugins-elevenlabs"]
cartesia = ["livekit-plugins-cartesia"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=src",
    "--cov-report=html",
    "--cov-report=term-missing",
]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "**/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
